src/backend/broker/requirements.txt
django-cors-headers
django-crontab
django-redis
mysqlclient
redis
djangorestframework